requirements.txt
pydantic~=2.9.2
puzpy~=0.2.6
lxml~=5.3.0
unidecode~=1.3.8
pytest~=8.3.3
openai~=1.54.4
python-dotenv~=1.0.1